requirements.txt
selenium
python-dotenv
ttkbootstrap
SQLAlchemy
webdriver-manager